requirements.txt
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.32.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0
opencv-python-headless>=4.8.0
matplotlib>=3.7.0
realesrgan>=0.3.0
basicsr>=1.4.2
tqdm>=4.65.0
